GTK GUI extracted, console mode added.
